requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
python-dotenv
google-genai